Day_05-ORM/requirement.txt
asgiref==3.11.0
Django==6.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
python-decouple==3.8
